put ToastMatcher in own file and try something else for the test
