+++ app/src/androidTest/java/ch/epfl/sdp/ToastMatcher.kt
@@ -1,0 +1,34 @@
+package ch.epfl.sdp
+
+import android.os.IBinder
+import android.view.WindowManager
+import androidx.test.espresso.Espresso
+import androidx.test.espresso.Root
+import androidx.test.espresso.assertion.ViewAssertions
+import androidx.test.espresso.matcher.ViewMatchers
+import org.hamcrest.Description
+import org.hamcrest.TypeSafeMatcher
+
+
+/**
+ * class to test whether or not a Toast Message is displayed
+ */
+class ToastMatcher : TypeSafeMatcher<Root>() {
+    override fun describeTo(description: Description) {
+        description.appendText("is toast")
+    }
+
+    @Suppress("DEPRECATION")
+    public override fun matchesSafely(root: Root): Boolean {
+        val type: Int = root.windowLayoutParams.get().type
+        if (type == WindowManager.LayoutParams.TYPE_TOAST) {
+            val windowToken: IBinder = root.decorView.windowToken
+            val appToken: IBinder = root.decorView.applicationWindowToken
+            if (windowToken === appToken) { // windowToken == appToken means this window isn't contained by any other windows.
+                // if it was a window for an activity, it would have TYPE_BASE_APPLICATION.
+                return true
+            }
+        }
+        return false
+    }
+}
